requirements.txt
flask
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
imaplib2
beautifulsoup4
email-validator
